Refactor CartPrice to use BigDecimal for storing price values instead of Int

+++ core/domain/src/main/java/ru/livetyping/zarina/core/domain/model/cart/CartPrice.kt
@@ -1,10 +1,12 @@
 package ru.livetyping.zarina.core.domain.model.cart
 
+import java.math.BigDecimal
+
 // Marked as stable on config/compose/stability_config.txt
 public data class CartPrice(
-    val cartPrice: Int,
-    val discountSize: Int,
-    val finalPrice: Int,
-    val deliveryPrice: Int?,
-    val giftCertificateRedemptionValue: Int?,
+    val cartPrice: BigDecimal,
+    val discountSize: BigDecimal,
+    val finalPrice: BigDecimal,
+    val deliveryPrice: BigDecimal?,
+    val giftCertificateRedemptionValue: BigDecimal?,
 )
